+++ idearium-lib/index.js
@@ -3,13 +3,14 @@
 module.exports = {
 
     // Packages here.
-    mq: require('./lib/mq'),
+    emailServices: require('./lib/email-services'),
+    kue: require('./lib/kue'),
     logs: require('./lib/logs'),
+    middleware: require('./middleware'),
+    mq: require('./lib/mq'),
     query: require('./lib/query'),
     util: require('./lib/util'),
     utils: require('./lib/utils'),
-    middleware: require('./middleware'),
-    emailServices: require('./lib/email-services'),
 
     // Export classes here.
     Config: require('./lib/config'),