+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is a history of the changes made to idearium-lib.
 
 - Simplified `common/exception` now that Opbeat should become standard practice.
 - Added a `query` file to help create performant queries.
+- Added `kue` lib and common files.
 
 ## 1.0.0-alpha.13